+++ domain/lemmy/pagination/src/commonMain/kotlin/com/diegoberaldin/raccoonforlemmy/domain/lemmy/pagination/DefaultMultiCommunityPaginator.kt
@@ -1,11 +1,48 @@
-package com.github.diegoberaldin.raccoonforlemmy.unit.multicommunity.utils
+package com.diegoberaldin.raccoonforlemmy.domain.lemmy.pagination
 
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.ListingType
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.PostModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.SortType
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.repository.PostRepository
 
-internal class CommunityPaginator(
+internal class DefaultMultiCommunityPaginator(
+    private val postRepository: PostRepository,
+) : MultiCommunityPaginator {
+    private var paginators = emptyList<Paginator>()
+
+    override val canFetchMore: Boolean
+        get() = paginators.any { it.canFetchMore }
+
+    override fun setCommunities(ids: List<Long>) {
+        paginators = ids.map {
+            Paginator(
+                communityId = it,
+                postRepository = postRepository,
+            )
+        }
+    }
+
+    override fun reset() {
+        paginators.forEach { it.reset() }
+    }
+
+    override suspend fun loadNextPage(
+        auth: String?,
+        sort: SortType,
+    ): List<PostModel> = buildList {
+        for (paginator in paginators) {
+            if (paginator.canFetchMore) {
+                val elements = paginator.loadNextPage(
+                    auth = auth,
+                    sort = sort,
+                )
+                addAll(elements)
+            }
+        }
+    }.sortedByDescending { it.publishDate }
+}
+
+private class Paginator(
     private val communityId: Long,
     private val postRepository: PostRepository,
 ) {
@@ -23,7 +60,6 @@ internal class CommunityPaginator(
     suspend fun loadNextPage(
         auth: String?,
         sort: SortType,
-        currentIds: List<Long>,
     ): List<PostModel> {
         val (result, nextPage) = postRepository.getAll(
             auth = auth,
@@ -33,15 +69,7 @@ internal class CommunityPaginator(
             type = ListingType.All,
             sort = sort,
             communityId = communityId,
-        )?.let {
-            // prevents accidental duplication
-            val posts = it.first
-            it.copy(
-                first = posts.filter { p1 ->
-                    p1.id !in currentIds
-                },
-            )
-        } ?: (null to null)
+        ) ?: (null to null)
         if (!result.isNullOrEmpty()) {
             currentPage++
         }
